♻️  refactoring code and fix package.json
