--- packages/lisk-passphrase/.gitignore
@@ -1,3 +1,0 @@
-# Compiled files
-src/**/*.js
-test/**/*.js
