--- packages/lisk-passphrase/src/index.ts
@@ -1,21 +1,0 @@
-/*
- * Copyright © 2018 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- *
- */
-import * as Mnemonic from 'bip39';
-import * as validation from './validation';
-
-export = {
-	Mnemonic,
-	validation,
-};
